% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abort.R
\name{abort}
\alias{abort}
\alias{inform}
\alias{warn}
\title{Signal an error, warning, or message with cli formatting}
\usage{
abort()

inform()

warn()
}
\arguments{
\item{x}{\code{character}.}
}
\value{
Console output, with invisible return of \code{x} input.
}
\description{
Signal an error, warning, or message with cli formatting
}
\note{
Updated 2021-07-12.
}
\examples{
x <- "hello world"
try({
    abort(x)
})
try({
    warn(x)
})
inform(x)
}
\seealso{
\itemize{
\item \code{\link[cli:cli_abort]{cli::cli_abort()}}.
\item \code{\link[cli:cli_abort]{cli::cli_inform()}}.
\item \code{\link[cli:cli_abort]{cli::cli_warn()}}.
}
}
