% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{AcidCLI-package}
\alias{AcidCLI}
\alias{AcidCLI-package}
\title{AcidCLI}
\description{
Command line interface helpers for Acid Genomics packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.acidgenomics.com/packages/acidcli/}
  \item \url{https://github.com/acidgenomics/r-acidcli/}
  \item Report bugs at \url{https://github.com/acidgenomics/r-acidcli/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
