% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{alert}
\alias{alert}
\alias{alertInfo}
\alias{alertSuccess}
\alias{alertWarning}
\title{Allert messages}
\usage{
alert(x)

alertInfo(x)

alertSuccess(x)

alertWarning(x)
}
\description{
Allert messages
}
\note{
Updated 2021-01-05.
}
\seealso{
\itemize{
\item \code{\link[cli:cli_alert]{cli::cli_alert()}}.
\item \code{\link[cli:cli_alert]{cli::cli_alert_info()}}.
\item \code{\link[cli:cli_alert]{cli::cli_alert_success()}}.
\item \code{\link[cli:cli_alert]{cli::cli_alert_warning()}}.
}
}
