% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{header}
\alias{header}
\alias{h1}
\alias{h2}
\alias{h3}
\alias{h4}
\alias{h5}
\alias{h6}
\alias{h7}
\title{Header level}
\usage{
h1(x)

h2(x)

h3(x)

h4(x)

h5(x)

h6(x)

h7(x)
}
\arguments{
\item{x}{\code{character}.
Text to display as header.}
}
\description{
Header level
}
\note{
Updated 2021-01-04.
}
\examples{
h1("Level 1")
h2("Level 2")
}
\seealso{
\itemize{
\item \code{\link[cli:cli_h1]{cli::cli_h1()}}, \code{\link[cli:cli_h1]{cli::cli_h2()}}, \code{\link[cli:cli_h1]{cli::cli_h3()}}.
}
}
