% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list}
\alias{list}
\alias{dl}
\alias{ol}
\alias{ul}
\title{Lists}
\usage{
dl(x)

ol(x)

ul(x)
}
\arguments{
\item{x}{\code{character(1)}.
Text to be displayed as list items.}
}
\description{
Lists
}
\note{
Updated 2021-01-04.
}
\examples{
dl(c(a = "AAA", b = "BBB"))
ul(c("Item 1", "Item 2"))
}
