% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{blue}
\alias{bold}
\alias{cyan}
\alias{green}
\alias{magenta}
\alias{red}
\alias{underline}
\alias{yellow}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{crayon}{\code{\link[crayon:blue]{blue()}}, \code{\link[crayon:bold]{bold()}}, \code{\link[crayon:cyan]{cyan()}}, \code{\link[crayon:green]{green()}}, \code{\link[crayon:magenta]{magenta()}}, \code{\link[crayon:red]{red()}}, \code{\link[crayon:underline]{underline()}}, \code{\link[crayon:yellow]{yellow()}}}
}}

