% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{text}
\alias{text}
\alias{txt}
\alias{verbatim}
\title{CLI text}
\usage{
txt(x)

verbatim(x)
}
\arguments{
\item{x}{\code{character}.}
}
\value{
Console output, with invisible return of \code{x} input.
}
\description{
CLI text
}
\section{Functions}{
\itemize{
\item \code{txt}: Print to screen, wrapping to screen with automatically.
Exporting \code{txt} instead of \code{text} here, to avoid masking \code{stats::text}.

\item \code{verbatim}: Print to screen, without any automatic wrapping.
}}

\note{
Updated 2020-01-05.
}
\examples{
## Wrapping
txt(c("hello", " world"))

## No wrapping
verbatim(c("hello", "world"))
}
