% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline.R
\name{toVarString}
\alias{toVarString}
\title{Convert input to inline variable string}
\usage{
toVarString(x, n = Inf, class = "var")
}
\arguments{
\item{x}{\code{character}.}

\item{n}{\code{integer(1)} or \code{Inf}.
Maximum number of elements to include.}
}
\value{
\code{character(1)}.
}
\description{
Convert input to inline variable string
}
\note{
Updated 2021-01-10.
}
\examples{
x <- c("aaa", "bbb", "ccc", "ddd")
x <- toVarString(x, n = 3L)
x
}
