% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abort.R
\name{abort}
\alias{abort}
\alias{inform}
\alias{warn}
\title{Signal an error, warning, or message with cli formatting}
\usage{
abort(x, call = TRUE)

inform(x)

warn(x)
}
\arguments{
\item{x}{\code{character} or \code{condition} (i.e. returned from \code{stop} or \code{warning}).}

\item{call}{\code{logical(1)}.
Include call information in the return message.}
}
\value{
Console output, with invisible return of \code{x} input.
}
\description{
Signal an error, warning, or message with cli formatting
}
\note{
Updated 2023-01-30.
}
\examples{
x <- "hello world"
try({
    abort(x)
})
try({
    warn(x)
})
inform(x)
}
\seealso{
\itemize{
\item \code{cli::cli_abort()}.
\item \code{cli::cli_inform()}.
\item \code{cli::cli_warn()}.
\item \code{simpleCondition()}.
\item \code{simpleError()}.
\item \code{simpleWarning()}.
\item \code{simpleMessage()}.
}
}
