% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abort.R
\name{abort}
\alias{abort}
\alias{inform}
\alias{warn}
\title{Signal an error, warning, or message with cli formatting}
\usage{
abort(x, .call = TRUE)

inform(x)

warn(x)
}
\arguments{
\item{x}{\code{character} or \code{condition} (i.e. returned from \code{stop} or \code{warning}).}

\item{`.call`}{\code{logical(1)}.
Include call information in the return message.}
}
\value{
Console output, with invisible return of \code{x} input.
}
\description{
Signal an error, warning, or message with cli formatting
}
\note{
Updated 2021-08-13.
}
\examples{
x <- "hello world"
try({
    abort(x)
})
try({
    warn(x)
})
inform(x)
}
\seealso{
\itemize{
\item \code{\link[cli:cli_abort]{cli::cli_abort()}}, \code{\link[cli:cli_abort]{cli::cli_inform()}}, \code{\link[cli:cli_abort]{cli::cli_warn()}}.
\item \code{simpleCondition()}, \code{simpleError()}, \code{simpleWarning()}, \code{simpleMessage()}.
}
}
