% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{alert}
\alias{alert}
\alias{alertDanger}
\alias{alertInfo}
\alias{alertSuccess}
\alias{alertWarning}
\title{Alert messages}
\usage{
alert(x)

alertDanger(x)

alertInfo(x)

alertSuccess(x)

alertWarning(x)
}
\arguments{
\item{x}{\code{character}.}
}
\value{
Console output, with invisible return of \code{x} input.
}
\description{
Alert messages
}
\note{
Updated 2021-08-18.
}
\examples{
x <- "hello world"
alert(x)
alertDanger(x)
alertInfo(x)
alertSuccess(x)
alertWarning(x)
}
\seealso{
\itemize{
\item \code{cli::cli_alert()}.
\item \code{cli::cli_alert_info()}.
\item \code{cli::cli_alert_success()}.
\item \code{cli::cli_alert_warning()}.
}
}
