% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list}
\alias{list}
\alias{dl}
\alias{ol}
\alias{ul}
\title{Lists}
\usage{
dl(x)

ol(x)

ul(x)
}
\arguments{
\item{x}{\code{character}.}
}
\value{
Console output, with invisible return of \code{x} input.
}
\description{
Lists
}
\note{
Updated 2021-01-18.
}
\examples{
dl(c(a = "AAA", b = "BBB"))
ol(c("Item 1", "Item 2"))
ul(c("Item 1", "Item 2"))
}
