% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{blue}
\alias{bold}
\alias{cyan}
\alias{green}
\alias{magenta}
\alias{red}
\alias{underline}
\alias{yellow}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{crayon}{\code{\link[crayon:crayon]{blue}}, \code{\link[crayon:crayon]{bold}}, \code{\link[crayon:crayon]{cyan}}, \code{\link[crayon:crayon]{green}}, \code{\link[crayon:crayon]{magenta}}, \code{\link[crayon:crayon]{red}}, \code{\link[crayon:crayon]{underline}}, \code{\link[crayon:crayon]{yellow}}}
}}

